code refactor to fix cyclic imports
