+++ api_app/models.py
@@ -1,7 +1,7 @@
 from django.db import models
 from django.contrib.postgres import fields as postgres_fields
 
-from api_app import utilities
+from api_app import helpers
 from .exceptions import AnalyzerRunException
 
 STATUS = [
@@ -47,7 +47,7 @@ class Job(models.Model):
     errors = postgres_fields.ArrayField(
         models.CharField(max_length=900), blank=True, default=list, null=True
     )
-    file = models.FileField(blank=True, upload_to=utilities.file_directory_path)
+    file = models.FileField(blank=True, upload_to=helpers.file_directory_path)
     tags = models.ManyToManyField(Tag, related_name="jobs", blank=True)
 
     @classmethod
@@ -65,5 +65,4 @@ class Job(models.Model):
     def __str__(self):
         if self.is_sample:
             return f'Job("{self.file_name}")'
-        else:
-            return f'Job("{self.observable_name}")'
+        return f'Job("{self.observable_name}")'
