+++ api_app/script_analyzers/file_analyzers/cuckoo_scan.py
@@ -4,14 +4,14 @@ import requests
 import logging
 
 from api_app.exceptions import AnalyzerRunException
-from api_app.utilities import get_binary
-from api_app.script_analyzers import classes
+from api_app.helpers import get_binary
+from api_app.script_analyzers.classes import FileAnalyzer
 from intel_owl import secrets
 
 logger = logging.getLogger(__name__)
 
 
-class CuckooAnalysis(classes.FileAnalyzer):
+class CuckooAnalysis(FileAnalyzer):
     def set_config(self, additional_config_params):
         # cuckoo installation can be with or without the api_token
         # it depends on version and configuration
