+++ api_app/script_analyzers/file_analyzers/intezer_scan.py
@@ -4,14 +4,14 @@ import requests
 import logging
 
 from api_app.exceptions import AnalyzerRunException
-from api_app.script_analyzers import classes
-from api_app.utilities import get_now_date_only, get_binary
+from api_app.script_analyzers.classes import FileAnalyzer
+from api_app.helpers import get_now_date_only, get_binary
 from intel_owl import secrets
 
 logger = logging.getLogger(__name__)
 
 
-class IntezerScan(classes.FileAnalyzer):
+class IntezerScan(FileAnalyzer):
     base_url: str = "https://analyze.intezer.com/api/v2-0"
 
     def set_config(self, additional_config_params):
