+++ api_app/script_analyzers/file_analyzers/signature_info.py
@@ -5,12 +5,12 @@ from subprocess import Popen, DEVNULL, PIPE
 from celery.exceptions import SoftTimeLimitExceeded
 
 from api_app.exceptions import AnalyzerRunException
-from api_app.script_analyzers import classes
+from api_app.script_analyzers.classes import FileAnalyzer
 
 logger = logging.getLogger(__name__)
 
 
-class SignatureInfo(classes.FileAnalyzer):
+class SignatureInfo(FileAnalyzer):
     def run(self):
         p = None
         results = {
