+++ api_app/script_analyzers/file_analyzers/vt2_scan.py
@@ -4,7 +4,7 @@ import logging
 import requests
 
 from api_app.exceptions import AnalyzerRunException
-from api_app.utilities import get_binary
+from api_app.helpers import get_binary
 from api_app.script_analyzers.observable_analyzers import vt2_get
 from api_app.script_analyzers import classes
 from intel_owl import secrets
